- Changes to scraper process - Removed unused code
